+++ lib/govuk_publishing_components/presenters/subscription_links_helper.rb
@@ -1,14 +1,12 @@
 module GovukPublishingComponents
   module Presenters
     class SubscriptionLinksHelper
-      attr_reader :feed_box_id, :email_signup_link, :email_signup_link_data_attributes, :feed_link_box_value
+      attr_reader :email_signup_link, :email_signup_link_data_attributes
 
       def initialize(local_assigns)
         @local_assigns = local_assigns
-        @feed_box_id = "feed-reader-#{SecureRandom.hex(2)}"
         @email_signup_link = @local_assigns[:email_signup_link]
         @email_signup_link_data_attributes = @local_assigns[:email_signup_link_data_attributes]
-        @feed_link_box_value = @local_assigns[:feed_link_box_value]
       end
 
       def email_signup_link_text
@@ -24,7 +22,7 @@ module GovukPublishingComponents
       end
 
       def component_data_is_valid?
-        email_signup_link.present? || feed_link.present? || feed_link_box_value.present?
+        email_signup_link.present? || feed_link.present?
       end
 
       def tracking_is_present?
@@ -32,16 +30,11 @@ module GovukPublishingComponents
       end
 
       def feed_link
-        return "#" if feed_link_box_value
-
         @local_assigns[:feed_link]
       end
 
       def feed_link_data_attributes
-        data = @local_assigns[:feed_link_data_attributes] || {}
-        data[:controls] = feed_box_id if feed_link_box_value
-        data[:expanded] = "false" if feed_link_box_value
-        data
+        @local_assigns[:feed_link_data_attributes] || {}
       end
     end
   end