+++ spec/lib/govuk_publishing_components/presenters/subscription_links_helper_spec.rb
@@ -1,6 +1,5 @@
 RSpec.describe GovukPublishingComponents::Presenters::SubscriptionLinksHelper do
   email_signup_link_data_attributes = { test: "value" }
-  feed_link_box_value = 3
 
   after do
     I18n.locale = :en
@@ -10,10 +9,8 @@ RSpec.describe GovukPublishingComponents::Presenters::SubscriptionLinksHelper do
     it "sets default values correctly" do
       allow(SecureRandom).to receive(:hex).and_return("12")
       subs = described_class.new({})
-      expect(subs.feed_box_id).to eq("feed-reader-12")
       expect(subs.email_signup_link).to be_nil
       expect(subs.email_signup_link_data_attributes).to be_nil
-      expect(subs.feed_link_box_value).to be_nil
     end
 
     it "accepts a value for email_signup_link" do
@@ -24,11 +21,6 @@ RSpec.describe GovukPublishingComponents::Presenters::SubscriptionLinksHelper do
     it "accepts a value for email_signup_link_data_attributes" do
       subs = described_class.new({ email_signup_link_data_attributes: })
       expect(subs.email_signup_link_data_attributes).to eq(email_signup_link_data_attributes)
-    end
-
-    it "accepts a value for feed_link_box_value" do
-      subs = described_class.new({ feed_link_box_value: })
-      expect(subs.feed_link_box_value).to eq(feed_link_box_value)
     end
   end
 
@@ -82,9 +74,6 @@ RSpec.describe GovukPublishingComponents::Presenters::SubscriptionLinksHelper do
 
       subs = described_class.new({ feed_link: "test" })
       expect(subs.component_data_is_valid?).to be(true)
-
-      subs = described_class.new({ feed_link_box_value: "test" })
-      expect(subs.component_data_is_valid?).to be(true)
     end
   end
 
@@ -106,11 +95,6 @@ RSpec.describe GovukPublishingComponents::Presenters::SubscriptionLinksHelper do
       expect(subs.feed_link).to be_nil
     end
 
-    it "is set to # if there is a feed_link_box_value" do
-      subs = described_class.new({ feed_link_box_value: })
-      expect(subs.feed_link).to eq("#")
-    end
-
     it "returns feed_link if set" do
       subs = described_class.new({ feed_link: "/link" })
       expect(subs.feed_link).to eq("/link")
@@ -128,17 +112,6 @@ RSpec.describe GovukPublishingComponents::Presenters::SubscriptionLinksHelper do
     it "returns passed attributes" do
       subs = described_class.new({ feed_link_data_attributes: })
       expect(subs.feed_link_data_attributes).to eq(feed_link_data_attributes)
-    end
-
-    it "includes data controls if there is a feed link box" do
-      allow(SecureRandom).to receive(:hex).and_return("12")
-      expected = {
-        key: "value",
-        expanded: "false",
-        controls: "feed-reader-12",
-      }
-      subs = described_class.new({ feed_link_data_attributes:, feed_link_box_value: })
-      expect(subs.feed_link_data_attributes).to eq(expected)
     end
   end
 end