+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ## Unreleased
 
+* Remove subscription links feed link button ([PR #5745](https://github.com/alphagov/govuk_publishing_components/pull/5745))
 * Remove branding from subscription links component ([PR #5743](https://github.com/alphagov/govuk_publishing_components/pull/5743))
 
 ## 70.1.0